+++ CHANGELOG.md
@@ -3,4 +3,4 @@
 ## 0.1.0
 - fix checkFormat task to use the passed folders instead of `.` 
 - initial release
-- add default grinder tasks
+- add default Grinder tasks